Redo Paul Vixie's bitstring.h #defines as inline functions, using uint32_t max-bits instead of "int", and rename include/bitstring.h to include/bitstring.i.
Change the names of the fixed-length column-store get/set value
functions to match the other bitstring functions, and move them
from include/btree.i to include/bitstring.i.

--HG--
rename : src/include/bitstring.h => src/include/bitstring.i

+++ src/include/api.h
@@ -45,7 +45,7 @@ struct __wt_btree {
 	uint32_t key_gap;		/* Btree instantiated key gap */
 	WT_BUF   key_srch;		/* Search key buffer */
 
-	u_int	 bitcnt;		/* Fixed-length field size in bits */
+	uint8_t	 bitcnt;		/* Fixed-length field size in bits */
 
 	int btree_compare_int;		/* Integer keys */
 					/* Comparison function */
